+++ Box Game Implementation/Box Game Implementation/blue_box.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "box.h"
+
+class BlueBox :public Box {
+	double largest;
+public:
+	BlueBox(double weight) : Box(weight), largest(weight) {};
+	double calculateScore(double absorbed_weight) override;
+	double cantorPairing(double a, double b);
+};
