+++ Box Game Implementation/Box Game Implementation/box.cpp
@@ -1,0 +1,9 @@
+#include "box.h"
+
+double Box::getScore() {
+	return this->score;
+}
+
+double Box::getSmallest() {
+	return this->smallest;
+}
