+++ Box Game Implementation/Box Game Implementation/green_box.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "box.h"
+
+class GreenBox :public Box {
+public:
+	GreenBox(double weight) : Box(weight) {};
+	double calculateScore(double absorbed_weight) override;
+};
