+++ Box Game Implementation/Box Game Implementation/main_box.cpp
@@ -17,156 +17,10 @@
 * -When all input token weights have been used up, the game ends, and the player with highest score wins.
  */
 
-#include<iostream>
-#include<vector>
-#include<algorithm>
-
-using namespace std;
-
-class Box {
-protected:
-	double weight;
-	double total_weight;
-	vector<double> absorbed_weights;
-	double score;
-	double smallest;
-public:
-	Box(double weight) : weight(weight), smallest(weight), score(0.0), total_weight(0.0) {};
-
-	virtual double calculateScore(double absorbed_weight) { return 0.0; }
-
-	double getScore() {
-		return this->score;
-	}
-
-	double getSmallest() {
-		return this->smallest;
-	}
-};
-
-class BlueBox:public Box {
-	double largest;
-public:
-	BlueBox(double weight) : Box(weight), largest(weight) {};
-
-	double calculateScore(double absorbed_weight) override {
-		if (absorbed_weight < smallest) {
-			smallest = absorbed_weight;
-		}
-		if (absorbed_weight > largest) {
-			largest = absorbed_weight;
-		}
-
-		return cantorPairing(smallest, largest);
-	}
-
-	double cantorPairing(double a, double b) {
-		return ((a + b) * (a + b + 1) / 2) + b;
-	}
-};
-
-class GreenBox :public Box {
-public:
-	GreenBox(double weight) : Box(weight) {};
-
-	double calculateScore(double absorbed_weight) override {
-		int start = -3;
-		int divisor = 3;
-		absorbed_weights.push_back(absorbed_weight);
-		int vector_length = absorbed_weights.size();
-		if (vector_length < 3) {
-			start = 0;
-			divisor = vector_length;
-		}
-		for (int i = vector_length - start-1; i < vector_length; i++) {
-			score = score += absorbed_weights[i];
-		}
-		score = score / divisor;
-
-		if (absorbed_weight < smallest or smallest == 0.0) {
-			smallest = absorbed_weight;
-		}
-
-		return score * score;
-	}
-};
-
-class Player {
-	double score;
-public:
-	Player() : score(0.0) {};
-
-	void addScore(double box_score){
-		this->score+=box_score;
-	}
-
-	double getScore() {
-		return this->score;
-	}
-};
-
-void printScore(Player* A, Player* B, string state="Current") {
-	cout << state << " score below: " << endl;
-	cout << "Player A: " << A->getScore() << endl;
-	cout << "Player B: " << B->getScore() << endl << endl;
-}
-
-void play(vector<double> weights) {
-	void findWinner(Player*, Player*);
-	double GetLightestBoxScore(Box*, Box*, Box*, Box*, double);
-
-	GreenBox g1(0.0);
-	GreenBox g2(0.1);
-	BlueBox l1(0.2);
-	BlueBox l2(0.3);
-
-	Box* b1 = dynamic_cast<Box*>(&g1);
-	Box* b2 = dynamic_cast<Box*>(&g2);
-	Box* b3 = dynamic_cast<Box*>(&l1);
-	Box* b4 = dynamic_cast<Box*>(&l2);
-
-	Player A;
-	Player B;
-
-	while (!weights.empty()) {
-		for (int i = 0; i < 2 && !weights.empty(); i++) {
-			if (i == 0) {
-				A.addScore(GetLightestBoxScore(b1, b2, b3, b4, weights.front()));
-			}
-			else {
-				B.addScore(GetLightestBoxScore(b1, b2, b3, b4, weights.front()));
-			}
-			weights.erase(weights.begin());
-			printScore(&A, &B);
-		}
-	}
-	findWinner(&A, &B);
-}
-
-void findWinner(Player* A, Player* B) {
-	if (A->getScore() > B->getScore()) {
-		cout << "Player A wins" << endl;
-	}
-	else {
-		cout << "Player B wins" << endl;
-	}
-	printScore(A, B, "Final");
-}
-
-double GetLightestBoxScore(Box* a, Box* b, Box* c, Box* d, double weight) {
-	double lowest_weight = numeric_limits<double>::max();
-	Box* box_with_lowest_weight = nullptr;
-	for (Box* box : { a,b,c,d }) {
-		if (lowest_weight > box->getSmallest()) {
-			lowest_weight = box->getSmallest();
-			box_with_lowest_weight = box;
-		}
-	}
-	box_with_lowest_weight->calculateScore(weight);
-	return box_with_lowest_weight->getScore();
-}
+#include "game.h"
 
 int main() {
-	vector<double> weights = { 2,3,1.5 };
-	play(weights);
+	vector<double> weights = { 2,3,1.5, 6.5, 7,1,3.3 };
+	Game g;
+	g.play(weights);
 }
