+++ Box Game Implementation/Box Game Implementation/player.cpp
@@ -1,0 +1,9 @@
+#include "player.h"
+
+void Player::addScore(double box_score) {
+	this->score += box_score;
+}
+
+double Player::getScore() {
+	return this->score;
+}
